Restructure the Clearing System APIs
Merge Clearing and Liquidation into MarginAccount API.
Rename Builder to Product and move some function from Trading.
