+++ examples/create_product.py
@@ -11,9 +11,9 @@ PRIVATE_KEY = "0x926b0e772d87247fb08832e7fd55e528ae5997680713367a4786c92e7d90915
 
 def main():
     app = afp.AFP(AUTONITY_RPC_URL, afp.PrivateKeyAuthenticator(PRIVATE_KEY))
-    builder = app.Builder()
+    product_api = app.Product()
 
-    product = builder.create_product(
+    product = product_api.create_product(
         symbol="SDK-TEST-1",
         description="Test Product 1",
         oracle_address="0xd8A8C5A492Fc2448cFcF980218c0F7D2De4d6FB3",
@@ -34,8 +34,8 @@ def main():
     )
     pprint(product.model_dump())
 
-    builder.register_product(product)
-    print(builder.product_state(product.id))
+    product_api.register_product(product)
+    print(product_api.product_state(product.id))
 
 
 if __name__ == "__main__":
